Add component for name and rename additionalPropertiesRenderer
