+++ core-ui/src/components/Extensibility/components/metadataSchema.js
@@ -2,20 +2,21 @@ export const METADATA_SCHEMA = {
   properties: {
     name: {
       type: 'string',
+      widget: 'Name',
     },
     labels: {
       type: 'object',
       additionalProperties: {
         type: 'string',
       },
-      widget: 'AdditionalProperties',
+      widget: 'KeyValuePair',
     },
     annotations: {
       type: 'object',
       additionalProperties: {
         type: 'string',
       },
-      widget: 'AdditionalProperties',
+      widget: 'KeyValuePair',
     },
   },
   type: 'object',
