+++ core-ui/src/components/Extensibility/ds/widgets-form/index.js
@@ -15,7 +15,8 @@ import { validators } from '@ui-schema/ui-schema/Validators/validators';
 import { StringRenderer } from './StringRenderer';
 import { NumberRenderer } from './NumberRenderer';
 import { SwitchRenderer } from './SwitchRenderer';
-import { AdditionalPropertiesRenderer } from './AdditionalPropertiesRenderer';
+import { NameRenderer } from './NameRenderer';
+import { KeyValuePairRenderer } from './KeyValuePairRenderer';
 import { CollapsibleRenderer } from './CollapsibleRenderer';
 import { GenericList } from './GenericList';
 
@@ -64,7 +65,8 @@ export const widgets = {
     Card: CardRenderer,
     LabelBox,
     */
-    AdditionalProperties: AdditionalPropertiesRenderer,
+    Name: NameRenderer,
+    KeyValuePair: KeyValuePairRenderer,
     FormGroup: CollapsibleRenderer,
   },
 };
